+++ Assets/Plugins/Demigiant/DOTween/Modules/DOTweenModuleSprite.cs
@@ -1,4 +1,4 @@
-﻿// Author: Daniele Giardini - http://www.demigiant.com
+// Author: Daniele Giardini - http://www.demigiant.com
 // Created: 2018/07/13
 
 #if !DOTWEEN_NOSPRITES // MODULE_MARKER