+++ Assets/_Project/Scripts/Battle/EnemyIntentUI.cs
@@ -1,4 +1,4 @@
-﻿using UnityEngine;
+using UnityEngine;
 using MutationChess.Core;
 using TMPro;
 using UnityEngine.UI;