+++ Assets/_Project/Scripts/Effects/FrostCardEnergyRefundEffect.cs
@@ -1,4 +1,4 @@
-﻿using UnityEngine;
+using UnityEngine;
 using MutationChess.UI;
 using MutationChess.Battle;