+++ Assets/_Project/Scripts/Effects/GiftEffect.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using UnityEngine;
 using MutationChess.UI;
 using MutationChess.Battle;