+++ Assets/_Project/Scripts/Effects/SlimeEffect.cs
@@ -1,4 +1,4 @@
-﻿using UnityEngine;
+using UnityEngine;
 using MutationChess.Core;
 using MutationChess.UI;
 using System.Collections.Generic;