+++ Assets/_Project/Scripts/Map/Data/NodeType.cs
@@ -1,4 +1,4 @@
-﻿namespace MutationChess.Map
+namespace MutationChess.Map
 {
     public enum NodeType
     {