+++ Assets/_Project/Scripts/UI/HandManager.cs
@@ -1,4 +1,4 @@
-﻿using DG.Tweening;
+using DG.Tweening;
 using MutationChess.Battle;
 using MutationChess.Core;
 using System.Collections;