+++ Packages/com.merry-yellow.code-assist/Editor/Preferences/RegistryMonitor.cs
@@ -1,4 +1,4 @@
-﻿
+
 /*
  * Thanks to gr0ss for the inspiration.
  *